Downsized the Grid Info Message to Use int16 instead of int32 for Position Data. Position is in CM that means Grid size is Limited to 650 Meters.

+++ src/drivers/uwb/rddrone/uwb_r4.h
@@ -84,7 +84,7 @@ typedef struct {  //needs higher accuracy?
 } gps_pos_t;
 
 typedef struct {
-	int32_t x, y, z; //axis in cm
+	int16_t x, y, z; //axis in cm
 } position_t; // Position of a device or target in 3D space
 
 typedef union {
